control-plane/migration-operator/requirements.txt
kopf~=1.35.1
kubernetes>=18.20.0
PyYAML~=6.0
requests~=2.26.0